Finish preliminary application rewrite
It compiles and runs now identically to the tutorial application, with the
exception of the details pane, which I completely removed.

Next steps are to continue the rewrite more invasively, ripping out the guts of
the tutorial class with a specific eye to improving overall const-correctness,
lifetime awareness, and exception safety, as well as patching up some of the
more troublesome and user-unfriendly OGRE APIs (e.g. ColourValue, String).